requirements.txt
pyipng~=1.0.3
pillow~=9.3.0
setuptools~=65.6.3
cython~=3.0.2
zlib~=1.2.13